@siying 2017-12-18
Please mention in HISTORY.md this improvement and the behavior change.

I still feel like we should give users a way to preserve the previous behavior. Can we do something like that:
```
if (read_options_.readahead_size == 0 || read_options_.readahead_size >= 8 * 1024) {
  // the prefetch logic
  ......
}
```