@sagar0 2018-01-19
Since we don't have readrandomwhilescanning benchmark implemented yet, I just tested using readrandom and made sure that there is no regression for point lookups. 

```
TEST_TMPDIR=/data/users/$USER/benchmarks/iter ./db_bench -benchmarks=readrandom -duration=60 -num=1000000000 -use_existing_db -statistics -histogram
```
Without this patch:
```
readrandom   :     303.736 micros/op 3289 ops/sec;    0.4 MB/s (194999 of 194999 found)
```
With this patch:
```
readrandom   :     301.907 micros/op 3312 ops/sec;     0.4 MB/s (198999 of 198999 found)
```